Allow all handler methods to have a field filter
